refactor: 메인, 전체 포스트, 카테고리별 포스트 목록 페이지 스타일 수정 및 공용 컴포넌트 분리 (#6)
