+++ frontend/src/common/components/CategorySection.tsx
