+++ frontend/src/components/main/CategoryLatestPosts.tsx
@@ -2,16 +2,22 @@ import React, { useEffect, useState } from "react";
 import styled from "styled-components";
 import { useCategory } from "@components/main/HomePage";
 import { fetchLatestPostsByCategoryId, Post } from "./postApi";
-import HorizontalPostCard from "./HorizontalPostCard";
+import HorizontalPostCard from "../../common/components/HorizontalPostCard";
 import { fetchCategories } from "../category/categoryApi";
 import { useNavigate } from "react-router-dom";
-import CategorySection from "./CategorySection";
+import CategorySection from "../../common/components/CategorySection";
+import DesktopCategoryWrapper from "../../common/components/DesktopCategoryWrapper";
+import FaBarsWrapper from "../../common/components/FaBarsWrapper";
+import Category from "@components/category/Category";
 import MoreButton from "@common/components/MoreButton";
 
 const CategoryLatestPosts: React.FC = () => {
   const { selectedCategoryId, setSelectedCategoryId } = useCategory(); // 카테고리 상태 공유
   const [posts, setPosts] = useState<Post[]>([]);
   const [categories, setCategories] = useState<{ categoryId: string; name: string }[]>([]);
+  const [isSticky, setIsSticky] = useState(false);
+  const [rightOffset, setRightOffset] = useState(0);
+  const [initialRightOffset, setInitialRightOffset] = useState(0);
   const navigate = useNavigate();
 
   // 최상단 카테고리를 기본 선택
@@ -67,17 +73,48 @@ const CategoryLatestPosts: React.FC = () => {
   const selectedCategoryName =
     categories.find((cat) => cat.categoryId === selectedCategoryId)?.name || "Category";
 
+  useEffect(() => {
+    const contentRow = document.getElementById("content-row");
+
+    const handleScroll = () => {
+      const contentRowRect = contentRow?.getBoundingClientRect();
+      const viewportWidth = window.innerWidth;
+
+      if (contentRowRect) {
+        const initialRightSpace = viewportWidth - contentRowRect.right;
+
+        setInitialRightOffset(initialRightSpace + 14.5);
+        setRightOffset(isSticky ? initialRightSpace + 30 : 0);
+        setIsSticky(contentRowRect.top <= -200);
+      }
+    };
+
+    handleScroll();
+    window.addEventListener("scroll", handleScroll);
+
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, [isSticky]);
+
   return (
     <Container>
-      <Header>
-        <CategorySection title={selectedCategoryName} />
+      <CategorySection title={selectedCategoryName} small />
+      <MoreButtonContainer>
         <MoreButton onClick={handleMoreButtonClick} />
-      </Header>
-      <PostsGrid>
-        {posts.map((post) => (
-          <HorizontalPostCard key={post.postId} post={post} />
-        ))}
-      </PostsGrid>
+      </MoreButtonContainer>
+      <ContentWrapper id="content-row">
+        <HorizontalPostGrid>
+          {posts.map((post) => (
+            <HorizontalPostCard key={post.postId} post={post} />
+          ))}
+        </HorizontalPostGrid>
+        <DesktopCategoryWrapper marginTop={0}>
+          <Category />
+        </DesktopCategoryWrapper>
+        <FaBarsWrapper
+          isSticky={isSticky}
+          rightOffset={isSticky ? rightOffset : initialRightOffset}
+        />
+      </ContentWrapper>
     </Container>
   );
 };
@@ -93,25 +130,29 @@ const Container = styled.div`
   }
 `;
 
-const Header = styled.div`
+const ContentWrapper = styled.div`
   display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 1rem;
-  justify-content: space-between; /* 카테고리 제목을 왼쪽에 정렬하고 버튼을 오른쪽으로 */
-  gap: 1rem; /* CategorySection과 MoreButton 사이의 간격 추가 */
-`;
-
-// PostsGrid에 포함된 각 카드의 정렬 및 레이아웃 정의
-const PostsGrid = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-  max-width: 100%;
+  margin-top: 80px;
+  gap: 2rem;
 
   @media (max-width: 768px) {
-    padding: 0 1rem; /* 모바일에서는 좌우 여백 추가 */
+    flex-direction: column;
+    gap: 1rem;
   }
+`;
+
+const MoreButtonContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  margin-top: 1rem;
+`;
+
+const HorizontalPostGrid = styled.div`
+  width: 75%;
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+  margin-right: 50px;
 `;
 
 export default CategoryLatestPosts;
